requirements.txt
shioaji
pandas
matplotlib
python-dotenv
BeautifulSoup4
seaborn